day24/streamsocial-headers/requirements.txt
kafka-python==2.0.2
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.1
asyncio-mqtt==0.16.1
redis==5.0.1
prometheus-client==0.19.0
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-kafka-python==0.43b0
jinja2==3.1.3
aiofiles==23.2.1
websockets==12.0
pytest==8.0.0
pytest-asyncio==0.23.5
docker==7.0.0
confluent-kafka==2.3.0
avro==1.11.3
